Migrate arrow parsing logic to dedicated utils module
